refactor(tests): move resource pool parallel test to api-db
