+++ crates/api-db/src/resource_pool.rs
@@ -1482,13 +1482,97 @@ pub async fn create_common_pools(
 
 #[cfg(test)]
 mod tests {
+    use std::str::FromStr;
+
     use carbide_instrument::MetricKind;
     use carbide_instrument::testing::{MetricsCapture, capture_logs, capture_logs_async};
     use carbide_test_support::Outcome::*;
     use carbide_test_support::query_counter::count_queries;
     use carbide_test_support::{Case, Check, check_cases, check_values};
+    use sqlx::postgres::{PgConnectOptions, PgPoolOptions};
 
     use super::*;
+
+    #[tokio::test(flavor = "multi_thread", worker_threads = 50)]
+    async fn test_parallel() -> Result<(), eyre::Report> {
+        // We can't use #[sqlx::test] here because we need a multi-threaded
+        // executor with 50 worker threads. Instead we manage the test database
+        // lifecycle manually, using a random name so multiple test runs (or
+        // parallel CI jobs) never collide on the same Postgres instance.
+        let base_url = std::env::var("DATABASE_URL")?;
+        // ResourcePool.name is varchar(32), so keep the DB name short.
+        let short_id = &uuid::Uuid::new_v4().simple().to_string()[..8];
+        let db_name = format!("test_par_{short_id}");
+        let base_options = PgConnectOptions::from_str(&base_url)?;
+
+        let admin = PgPoolOptions::new()
+            .connect_with(base_options.clone())
+            .await?;
+
+        sqlx::query(sqlx::AssertSqlSafe(format!(
+            "CREATE DATABASE \"{db_name}\""
+        )))
+        .execute(&admin)
+        .await?;
+        let db_pool = PgPoolOptions::new()
+            .connect_with(base_options.database(&db_name))
+            .await?;
+        crate::migrations::migrate(&db_pool).await?;
+
+        let mut txn = db_pool.begin().await?;
+        let pool = Arc::new(ResourcePool::new(db_name.clone(), ValueType::Integer));
+
+        crate::resource_pool::populate(
+            &pool,
+            &mut txn,
+            (1..=5_000).map(|i| i.to_string()).collect(),
+            true,
+        )
+        .await?;
+        txn.commit().await?;
+
+        let mut handles = Vec::with_capacity(50);
+        let all_values = Arc::new(tokio::sync::Mutex::new(HashSet::new()));
+        for i in 0..50 {
+            let all_values = all_values.clone();
+            let p = pool.clone();
+            let db_pool_c = db_pool.clone();
+            let handle = tokio::task::spawn(async move {
+                let mut got = Vec::with_capacity(100);
+                for _ in 0..100 {
+                    let mut txn = db_pool_c.begin().await.unwrap();
+                    got.push(
+                        crate::resource_pool::allocate(
+                            &p,
+                            &mut txn,
+                            OwnerType::Machine,
+                            &i.to_string(),
+                            None,
+                        )
+                        .await
+                        .unwrap(),
+                    );
+                    txn.commit().await.unwrap();
+                }
+                all_values.lock().await.extend(got.clone());
+            });
+            handles.push(handle);
+        }
+        futures::future::join_all(handles).await;
+        drop(pool);
+        db_pool.close().await;
+
+        assert_eq!(all_values.lock().await.len(), 5_000);
+
+        // WITH (FORCE) terminates any lingering backends before dropping,
+        // avoiding the flaky "database is being accessed by other users" error.
+        let drop_stmt = format!("DROP DATABASE \"{db_name}\" WITH (FORCE)");
+        sqlx::query(sqlx::AssertSqlSafe(drop_stmt))
+            .execute(&admin)
+            .await?;
+        admin.close().await;
+        Ok(())
+    }
 
     const RESOURCE_POOL_LIFECYCLE_FAILURES_METRIC: &str =
         "carbide_resource_pool_lifecycle_failures_total";
